Use single, async subscription for all topics
Instead of opening a new subscription for each topic, open a single
subcription for all topics and route them with an if-elif structure
like a rail yard. Additionally, make each message run inside its own
task so that the message queue is never busy. Use a TaskGroup in order
to keep a strong reference to the tasks, otherwise they get erased by
the GC.

Fix #7