Fix bug that makes recon service hang on error
Service would hang upon error during message coordination in main
because the task would exit and the cancel event would be set in the
foreground task, making it impossible for the event loop to stop
waiting on the event.

Also, major refactor to make the code cleaner and extract messaging
logic from  main.